+++ tests/fetch_all.py
@@ -1,0 +1,4 @@
+import serverjars
+
+all = serverjars.fetchAll(category='snapshot', max=10)
+print(all)