+++ tests/fetch_details.py
@@ -1,0 +1,4 @@
+import serverjars
+
+details = serverjars.fetchDetails('paper', '1.19.3')
+print(details)