+++ tests/fetch_latest.py
@@ -1,0 +1,4 @@
+import serverjars
+
+latest = serverjars.fetchLatest(category='paper')
+print(latest)