+++ tests/fetch_subtypes.py
@@ -1,0 +1,4 @@
+import serverjars
+
+subtypes = serverjars.fetchTypes(type='bedrock')
+print(subtypes)