+++ tests/fetch_types.py
@@ -1,0 +1,4 @@
+import serverjars
+
+types = serverjars.fetchTypes()
+print(types)