+++ CHANGELOG.md
@@ -1,8 +1,13 @@
-# Changelog
+## CHANGELOG v1.1.0
+- Breaking Change! package has been renamed from `ServerJars` to `serverjars`.
+- constants are no longer enums.
+- _type now uses fetchTypes instead of a data set.
+- Improved method docs.
+- Added `finishcommand` to downloadJar. Once jar has been downloaded it will call this function.
 
-## v1.0.1
+## CHANGELOG v1.0.1
 - Fixed `ServerJars.fetchTypes()`
 - Added example.py
 
-## v1.0.0
+## CHANGELOG v1.0.0
 - initial release